requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
langchain
langchain-groq
requests
python-dotenv